feature-policy/experimental-features/intrinsicsize-with-unsized-media.tentative.https.sub.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe></iframe>
<script>
'use strict';
const default_width = 300;
const default_height = 150;
// default ratio = default width (300) / default height (150)
const default_ratio = 2;
const srcs = [
  "/feature-policy/experimental-features/resources/image.jpg",
  "/feature-policy/experimental-features/resources/image.png",
  "/feature-policy/experimental-features/resources/image.svg",
  "/feature-policy/experimental-features/resources/image.bmp"
];
const test_cases = [
// Test when no size is specified, img/video is laid out by the default size.
{expected_width: default_width, expected_height: default_height, expected_naturalWidth: default_width, expected_naturalHeight: default_height},
{attribute: "intrinsicsize", value: "500 x 400", expected_width: 500, expected_height: 400, expected_naturalWidth: 500, expected_naturalHeight: 400},
// Test when only one dimension is specified, img/video uses the default length for
// the other dimension.
{attribute: "width", value: 500, expected_width: 500, expected_height: 500 * (1 / default_ratio), expected_naturalWidth: default_width, expected_naturalHeight: default_height},
{attribute: "width", attribute1: "intrinsicsize", value: 500, value1: "200 x 100", expected_width: 500, expected_height: 500 * (100 / 200), expected_naturalWidth: 200, expected_naturalHeight: 100},
// Test when only one dimension is specified by CSS style, img/video uses the
// default length for the other dimension.
{attribute: "style", value: "height:800px;", expected_width: 800 * default_ratio, expected_height: 800, expected_naturalWidth: default_width, expected_naturalHeight: default_height},
{attribute: "style", attribute1: "intrinsicsize", value: "height:800px;", value1: "200 x 100", expected_width: 800 * default_ratio, expected_height: 800, expected_naturalWidth: 200, expected_naturalHeight: 100},
// Test when the size of the image is specified, img/video is laid out by the
// specified size.
{attribute: "style", value: "width:500px;height:800px;", expected_width: 500, expected_height:800, expected_naturalWidth: default_width, expected_naturalHeight: default_height},
{attribute: "style", attribute1: "intrinsicsize", value: "width:500px;height:800px;", value1: "200 x 100", expected_width: 500, expected_height:800, expected_naturalWidth: 200, expected_naturalHeight: 100}
];

for (var test of test_cases) {
  for (var src of srcs) {
    async_test(t=> {
      var img = document.createElement('IMG');
      var expected_width = test.expected_width;
      var expected_height = test.expected_height;
      var expected_naturalWidth = test.expected_naturalWidth;
      var expected_naturalHeight = test.expected_naturalHeight;
      img.setAttribute("src", src);
      if (typeof test.attribute !== "undefined") {
        img.setAttribute(test.attribute, test.value);
      }
      if (typeof test.attribute1 !== "undefined") {
        img.setAttribute(test.attribute1, test.value1);
      }
      img.addEventListener('load', t.step_func(function() {
        assert_equals(img.width, expected_width, 'width');
        assert_equals(img.height, expected_height, 'height');
        assert_equals(img.naturalWidth, expected_naturalWidth, 'naturalWidth');
        assert_equals(img.naturalHeight, expected_naturalHeight, 'naturalHeight');
        t.done();
      }));
      document.body.appendChild(img);
    }, 'Test image with attribute ' + test.attribute + '=' + test.value +
       ' and attribute ' + test.attribute1 + '=' + test.value1 + ' on src ' + src);
  }
}
</script>
</body>
